src/requirements.txt
absl-py==0.11.0
albumentations==0.5.2
astroid==2.4.2
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
coloredlogs==15.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@05573d7480a43e7edc61ba6061082c81dd3cde76
flake8==3.8.4
future==0.18.2
fvcore==0.1.2.post20210115
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
humanfriendly==9.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
iopath==0.1.2
isort==5.7.0
jedi==0.17.2
joblib==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.0
parso==0.7.1
Pillow==8.1.0
pluggy==0.13.1
portalocker==2.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pydot==1.4.1
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
ranger-fm==1.9.3
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
